Added OTAU support via Bintray.com
This needs more work but is essentially proven out.

See the notes here: https://docs.platformio.org/en/latest/platforms/espressif32.html#over-the-air-ota-update

We receive a TTN packet on port 2 which triggers a Wifi connection via SSID 'DoeSLiverpool'
to check on Bintray.com for a new firmware release.

This is downloaded, updated to flash and we reset.

The firmware displays the current revision. '0' is debug firmware.
'1' and higher is release firmware.

Because we rejoin the LoRa network after a reset we seem to keep getting retries
of the port 2 packet (and so would enter an infinite loop if constantly resetting).

Instead we receive the retries port 2 message again, try to update firmware again,
and detect there is no update, at which time we sleep until another wake event. As
we haven't reset, the wake event causes a packet send without rejoining and the
downlink message is not resent.

TODO:

- support configurable SSID, PASSWORD in TTN port 2 message (currently hardcoded)
- support timeout and retry on Wifi AP connection
- support timeout and retry on firmware update check / download
- support sending of Mitems f/w revision

NOTE:

We use a dual partition schema, so have two alternate images in flash. The downloader
switches between them. You can find if debugging that your debug version updates,
then the "alternate" is running. So even if you flash a new debug image into the device
it won't run this code, instead it will run the alternate, which is confusing. Check
the output device firmware revision to check what code is running. To completely erase
the device back to a "known state" use

esptool.exe -p COMx erase_flash

Signed-off-by: Alex J Lennon <ajlennon@dynamicdevices.co.uk>